+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 ### Removed
 
 ### Fixed
+1. Restore unpickling of `ThermoRawLoader`s
 
 ### Security